Handle non data base urls schemes in loadHtml
Skip percent encodig in case of setHtml call with baseUrl
with non data scheme. See for reference:

https://developer.android.com/reference/android/webkit/WebView#
loadDataWithBaseURL(java.lang.String,%20java.lang.String,
%20java.lang.String,%20java.lang.String,%20java.lang.String)

Add test case to cover the issue, note this skips the
webenigne backend as it requires qmlengine to function.
This will be cleaned up in follow up patches.

Note the base url setting is not supported on "webview2".

Fixes: QTBUG-136082
Fixes: QTBUG-134723
Pick-to: 6.10 6.9 6.8
Change-Id: I6891ec9271fcf3ee78048a6b040d39b818087fa4
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>