Minibrowser: Make sure icons aren't rendered in black
The default Qt Quick style has changed, resulting in icons being
rendered in all black.

Qt Quick docs state: The icon color is specified by default so that it
matches the text color in different states. In order to use an icon with
the original colors, set the color to "transparent".

Pick-to: 6.10
Change-Id: I3a3c3494268708be05b209c98fb43f8afa0536e3
Reviewed-by: Michal Klocek <michal.klocek@qt.io>